requirements.txt
# 核心依赖
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# 机器学习
scikit-learn>=1.3.0

# 深度学习 (可选，根据需要安装)
# torch>=2.0.0
# tensorflow>=2.13.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 工具
tqdm>=4.65.0
joblib>=1.3.0
